moved searchFilter to Entity
