+++ src/AppBundle/Entity/Form/SearchFilterType.php
@@ -47,7 +47,7 @@ class SearchFilterType extends AbstractType
     {
         $resolver->setDefaults(array(
             "translation_domain" => "validators",
-            "data_class" => "AppBundle\Entity\Form\SearchFilter",
+            "data_class" => "AppBundle\Entity\SearchFilter",
             "csrf_protection" => true,
             "csrf_field_name" => "_token",
             // a unique key to help generate the secret token
